Re #1813 Added check for CFT.appliesTo field validity. Fixed MeveoCF test file.

+++ meveo-api/src/main/java/org/meveo/api/CustomFieldTemplateApi.java
@@ -1,6 +1,7 @@
 package org.meveo.api;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
@@ -23,7 +24,9 @@ import org.meveo.model.crm.custom.CustomFieldStorageTypeEnum;
 import org.meveo.model.crm.custom.CustomFieldTypeEnum;
 import org.meveo.service.catalog.impl.CalendarService;
 import org.meveo.service.crm.impl.CustomFieldTemplateService;
-
+import org.meveo.service.custom.CustomizedEntity;
+import org.meveo.service.custom.CustomizedEntityService;
+import org.meveo.util.EntityCustomizationUtils;
 
 /**
  * @author Edward P. Legaspi
@@ -35,7 +38,10 @@ public class CustomFieldTemplateApi extends BaseApi {
     private CalendarService calendarService;
 
     @Inject
-    private CustomFieldTemplateService customFieldTemplateService; 
+    private CustomFieldTemplateService customFieldTemplateService;
+
+    @Inject
+    private CustomizedEntityService customizedEntityService;
 
     public void create(CustomFieldTemplateDto postData, String appliesTo, User currentUser) throws MeveoApiException, BusinessException {
 
@@ -59,14 +65,15 @@ public class CustomFieldTemplateApi extends BaseApi {
         }
         if (postData.getFieldType() == CustomFieldTypeEnum.CHILD_ENTITY && (postData.getStorageType() != CustomFieldStorageTypeEnum.LIST || postData.isVersionable())) {
             throw new InvalidParameterException("Custom field of type CHILD_ENTITY only supports unversioned values and storage type of LIST");
-        }        
-        if (postData.getFieldType() == CustomFieldTypeEnum.CHILD_ENTITY && (postData.getChildEntityFieldsForSummary() == null || postData.getChildEntityFieldsForSummary().isEmpty())) {
+        }
+        if (postData.getFieldType() == CustomFieldTypeEnum.CHILD_ENTITY
+                && (postData.getChildEntityFieldsForSummary() == null || postData.getChildEntityFieldsForSummary().isEmpty())) {
             missingParameters.add("childEntityFieldsForSummary");
         }
         handleMissingParameters();
 
         if (appliesTo != null) {
-          postData.setAppliesTo(appliesTo);
+            postData.setAppliesTo(appliesTo);
 
         } else {
             // Support for old API
@@ -75,6 +82,10 @@ public class CustomFieldTemplateApi extends BaseApi {
             } else {
                 appliesTo = postData.getAppliesTo();
             }
+        }
+
+        if (!getCustomizedEntitiesAppliesTo(currentUser.getProvider()).contains(appliesTo)) {
+            throw new InvalidParameterException("appliesTo", appliesTo);
         }
 
         if (customFieldTemplateService.findByCodeAndAppliesToNoCache(postData.getCode(), appliesTo, currentUser.getProvider()) != null) {
@@ -108,8 +119,9 @@ public class CustomFieldTemplateApi extends BaseApi {
         }
         if (postData.getFieldType() == CustomFieldTypeEnum.CHILD_ENTITY && (postData.getStorageType() != CustomFieldStorageTypeEnum.LIST || postData.isVersionable())) {
             throw new InvalidParameterException("Custom field of type CHILD_ENTITY only supports unversioned values and storage type of LIST");
-        }      
-        if (postData.getFieldType() == CustomFieldTypeEnum.CHILD_ENTITY && (postData.getChildEntityFieldsForSummary() == null || postData.getChildEntityFieldsForSummary().isEmpty())) {
+        }
+        if (postData.getFieldType() == CustomFieldTypeEnum.CHILD_ENTITY
+                && (postData.getChildEntityFieldsForSummary() == null || postData.getChildEntityFieldsForSummary().isEmpty())) {
             missingParameters.add("childEntityFieldsForSummary");
         }
 
@@ -125,6 +137,10 @@ public class CustomFieldTemplateApi extends BaseApi {
             } else {
                 appliesTo = postData.getAppliesTo();
             }
+        }
+
+        if (!getCustomizedEntitiesAppliesTo(currentUser.getProvider()).contains(appliesTo)) {
+            throw new InvalidParameterException("appliesTo", appliesTo);
         }
 
         CustomFieldTemplate cft = customFieldTemplateService.findByCodeAndAppliesToNoCache(postData.getCode(), appliesTo, currentUser.getProvider());
@@ -148,6 +164,10 @@ public class CustomFieldTemplateApi extends BaseApi {
 
         handleMissingParameters();
 
+        if (!getCustomizedEntitiesAppliesTo(provider).contains(appliesTo)) {
+            throw new InvalidParameterException("appliesTo", appliesTo);
+        }
+
         CustomFieldTemplate cft = customFieldTemplateService.findByCodeAndAppliesTo(code, appliesTo, provider);
         if (cft != null) {
             customFieldTemplateService.remove(cft.getId());
@@ -165,6 +185,10 @@ public class CustomFieldTemplateApi extends BaseApi {
         }
 
         handleMissingParameters();
+
+        if (!getCustomizedEntitiesAppliesTo(provider).contains(appliesTo)) {
+            throw new InvalidParameterException("appliesTo", appliesTo);
+        }
 
         CustomFieldTemplate cft = customFieldTemplateService.findByCodeAndAppliesTo(code, appliesTo, provider);
 
@@ -282,6 +306,13 @@ public class CustomFieldTemplateApi extends BaseApi {
         }
         return cft;
     }
-    
-  
+
+    private List<String> getCustomizedEntitiesAppliesTo(Provider provider) {
+        List<String> cftAppliesto = new ArrayList<String>();
+        List<CustomizedEntity> entities = customizedEntityService.getCustomizedEntities(null, false, true, null, null, provider);
+        for (CustomizedEntity customizedEntity : entities) {
+            cftAppliesto.add(EntityCustomizationUtils.getAppliesTo(customizedEntity.getEntityClass(), customizedEntity.getEntityCode()));
+        }
+        return cftAppliesto;
+    }
 }
